Extract common logging classes to a common package
