+++ 260-delegate/src/main/java/io/harness/delegate/logging/DelegateStackdriverLogAppender.java
@@ -12,7 +12,7 @@ import static io.harness.network.SafeHttpCall.execute;
 import io.harness.annotations.dev.HarnessModule;
 import io.harness.annotations.dev.TargetModule;
 import io.harness.concurrent.HTimeLimiter;
-import io.harness.logging.AccessTokenBean;
+import io.harness.logging.common.AccessTokenBean;
 import io.harness.logging.RemoteStackdriverLogAppender;
 import io.harness.managerclient.DelegateAgentManagerClient;
 import io.harness.rest.RestResponse;
