+++ 420-delegate-service/src/test/java/io/harness/service/stackdriver/DelegateStackdriverLogServiceTest.java
@@ -18,7 +18,7 @@ import static org.mockito.Mockito.when;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.category.element.UnitTests;
 import io.harness.delegate.resources.DelegateStackDriverLog;
-import io.harness.logging.AccessTokenBean;
+import io.harness.logging.common.AccessTokenBean;
 import io.harness.logging.StackdriverLoggerFactory;
 import io.harness.ng.beans.PageRequest;
 import io.harness.rule.Owner;
