+++ 960-watcher/src/main/java/io/harness/managerclient/ManagerClientV2.java
@@ -9,7 +9,7 @@ package io.harness.managerclient;
 
 import io.harness.delegate.beans.DelegateConfiguration;
 import io.harness.delegate.beans.DelegateScripts;
-import io.harness.logging.AccessTokenBean;
+import io.harness.logging.common.AccessTokenBean;
 import io.harness.rest.RestResponse;
 
 import retrofit2.Call;
