+++ 980-commons/src/main/java/io/harness/logging/common/BUILD.bazel
@@ -1,0 +1,24 @@
+# Copyright 2023 Harness Inc. All rights reserved.
+# Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+# that can be found in the licenses directory at the root of this repository, also available at
+# https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+
+load("@rules_java//java:defs.bzl", "java_library")
+load("//:tools/bazel/macros.bzl", "run_analysis_per_module")
+
+java_library(
+    name = "module",
+    srcs = glob(["*.java"]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//:lombok",
+        "//980-commons/src/main/java/io/harness/logging/autolog:module",
+        "@maven//:ch_qos_logback_contrib_logback_jackson",
+        "@maven//:ch_qos_logback_contrib_logback_json_core",
+        "@maven//:ch_qos_logback_logback_classic",
+        "@maven//:ch_qos_logback_logback_core",
+        "@maven//:org_apache_commons_commons_lang3",
+    ],
+)
+
+run_analysis_per_module()
